+++ aqme/qdescp.py
@@ -18,17 +18,17 @@ from aqme.utils import (
     mol_from_sdf_or_mol_or_mol2,
     run_command
 )
-from aqme.qdescp_utils import get_boltz_avg_properties_xtb
-
-from aqme.crest import xyzall_2_xyz
-from aqme.xtb_to_json import (
+from aqme.qdescp_utils import (
+    get_boltz_avg_properties_xtb,
     read_fod,
     read_json,
     read_xtb,
     read_wbo,
     read_gfn1,
-    read_fukui,
+    read_fukui
 )
+
+from aqme.csearch_crest_utils import xyzall_2_xyz
 from scipy.spatial.distance import cdist
